pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "react-agent-compensation"
version = "0.1.0"
description = "Framework-agnostic compensation/rollback library for ReAct agents"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Kaviru Hapuarachchi" }
]
keywords = [
    "langchain",
    "agents",
    "compensation",
    "rollback",
    "recovery",
    "react-agent",
    "llm",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "pydantic>=2.0",
]

[project.optional-dependencies]
langchain = [
    "langchain>=0.3.0",
    "langgraph>=0.2.0",
]
mcp = [
    "langchain-mcp-adapters>=0.1.0",
    "mcp>=1.0.0",
    "fastmcp>=2.0.0",
]
llm = [
    "openai>=1.0.0",
    "anthropic>=0.20.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]
examples = [
    "langchain-google-genai>=2.0.0",
    "python-dotenv>=1.0.0",
    "langsmith>=0.1.0",
    "pymongo>=4.0.0",
]
all = [
    "react-agent-compensation[langchain,mcp,llm,dev,examples]",
]

[project.urls]
Homepage = "https://github.com/kaviruhapuarachchi/react-agent-compensation"
Repository = "https://github.com/kaviruhapuarachchi/react-agent-compensation"

[tool.hatch.build.targets.wheel]
packages = ["src/react_agent_compensation"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --tb=short"
